// libs/shared/api-client/src/generated/events/client/types.gen.ts
export interface Config<T extends ClientOptions = ClientOptions> extends Omit<CreateAxiosDefaults, 'auth' | 'baseURL' | 'headers' | 'method'>, CoreConfig {
    /**
     * Axios implementation. You can use this option to provide either an
     * `AxiosStatic` or an `AxiosInstance`.
     *
     * @default axios
     */
    axios?: AxiosStatic | AxiosInstance;
    /**
     * Base URL for all requests made by this client.
     */
    baseURL?: T['baseURL'];
    /**
     * An object containing any HTTP headers that you want to pre-populate your
     * `Headers` object with.
     *
     * {@link https://developer.mozilla.org/docs/Web/API/Headers/Headers#init See more}
     */
    headers?: AxiosRequestHeaders | Record<string, string | number | boolean | (string | number | boolean)[] | null | undefined | unknown>;
    /**
     * Throw an error instead of returning it in the response?
     *
     * @default false
     */
    throwOnError?: T['throwOnError'];
}

export interface RequestOptions<TData = unknown, ThrowOnError extends boolean = boolean, Url extends string = string>
    extends
// ...
export interface ClientOptions {
    baseURL?: string;
    throwOnError?: boolean;
}

export type RequestResult<TData = unknown, TError = unknown, ThrowOnError extends boolean = boolean> = ThrowOnError extends true
    ? Promise<AxiosResponse<TData extends Record<string, unknown> ? TData[keyof TData] : TData>>
    : Promise<
          | (AxiosResponse<TData extends Record<string, unknown> ? TData[keyof TData] : TData> & {
                error: undefined;
            })
          | (AxiosError<TError extends Record<string, unknown> ? TError[keyof TError] : TError> & {
                data: undefined;
                error: TError extends Record<string, unknown> ? TError[keyof TError] : TError;
            })
      >;

type MethodFn = <TData = unknown, TError = unknown, ThrowOnError extends boolean = false>(
    options: Omit<RequestOptions<TData, ThrowOnError>, 'method'>,
) => RequestResult<TData, TError, ThrowOnError>;

type SseFn = <
    TData = unknown,
    // eslint-disable-next-line @typescript-eslint/no-unused-vars
    _TError = unknown,
    ThrowOnError extends boolean = false,
>(
    options: Omit<RequestOptions<never, ThrowOnError>, 'method'>,
) => Promise<ServerSentEventsResult<TData>>;

type RequestFn = <TData = unknown, TError = unknown, ThrowOnError extends boolean = false>(
    options: Omit<RequestOptions<TData, ThrowOnError>, 'method'> & Pick<Required<RequestOptions<TData, ThrowOnError>>, 'method'>,
) => RequestResult<TData, TError, ThrowOnError>;

type BuildUrlFn = <
    TData extends {
        path?: Record<string, unknown>;
        query?: Record<string, unknown>;
        url: string;
    },
>(
    options: TData & Pick<RequestOptions<unknown, boolean>, 'axios' | 'baseURL' | 'paramsSerializer' | 'querySerializer'>,
) => string;

export type Client = CoreClient<RequestFn, Config, MethodFn, BuildUrlFn, SseFn> & {
    instance: AxiosInstance;
};